+++ ignite/templates/typed/placeholders.go
@@ -3,12 +3,10 @@ package typed
 //nolint:godot
 const (
 	Placeholder  = "// this line is used by starport scaffolding # 1"
-	Placeholder2 = "// this line is used by starport scaffolding # 2"
 	Placeholder3 = "// this line is used by starport scaffolding # 3"
 	Placeholder4 = "<!-- this line is used by starport scaffolding # 4 -->"
 
 	// Genesis
-	PlaceholderGenesisTypesImport   = "// this line is used by starport scaffolding # genesis/types/import"
 	PlaceholderGenesisTypesDefault  = "// this line is used by starport scaffolding # genesis/types/default"
 	PlaceholderGenesisTypesValidate = "// this line is used by starport scaffolding # genesis/types/validate"
 	PlaceholderGenesisModuleInit    = "// this line is used by starport scaffolding # genesis/module/init"